+++ src/contracts/index.ts
@@ -77,3 +77,4 @@ export * from "./PersonalWatchlistMapping";
 export * from "./PersonalMarketDataComposition";
 export * from "./PersonalMarketDataProviderCoverage";
 export * from "./PersonalMarketDataAlternativeProviderQualification";
+export * from "./PersonalMarketDataTwelveDataQualification";